fatto

+++ download mod/mod-vanilla-2/mod.py
@@ -4,19 +4,33 @@ import sys
 import requests
 import zipfile
 import down
+import filecmp
 
-global user
-user = os.getlogin()
-global agg
-MINECRAFT = 'C:/Users/'+user+'/AppData/Roaming/.minecraft/'
+global USER
+USER = os.getlogin()
+global agg, sc
+MINECRAFT = 'C:/users/'+USER+'/AppData/Roaming/.minecraft/'
+GITHUB = 'https://raw.githubusercontent.com/dj2828/Ultra_Vanilla_2/master/download%20mod/down/'
 agg=False
+sc=False
 
 try:
     def fine():
         os.system('cls')
         if agg:
+            shutil.rmtree('__pycache__/')
             try:
-                os.remove('modlist.json')
+                os.remove(MINECRAFT+'mods/modlist.json')
+            except:
+                pass
+            shutil.move('modlist.json', MINECRAFT+'mods/modlist.json')
+
+            print('\n\033[92mOra prova ad aprire minecarft 1.20.1 forge 47.4.0\033[0m')
+            input('')
+        elif sc:
+            shutil.rmtree('__pycache__/')
+            shutil.move('modlist.json', MINECRAFT+'mods/modlist.json')
+            try:
                 os.remove('forge.jar')
                 os.remove('forge-1.20.1-47.4.0-installer.jar.log')
             except:
@@ -27,7 +41,7 @@ try:
 
     def tx():
         print('\033[92mScaricamento texturepack\nAttendi...\033[0m')
-        response = requests.get('https://github.com/dj2828/sito/releases/download/mod_vanilla_2/Ultra-vanilla-2.zip')
+        response = requests.get(GITHUB+'Ultra-vanilla-2.zip')
         with open('Ultra-vanilla-2.zip', 'wb') as f:
             f.write(response.content)
         
@@ -45,7 +59,7 @@ try:
 
     def cose(a):
     # scarica cose
-        response = requests.get('https://github.com/dj2828/sito/releases/download/mod_vanilla_2/cose.zip')
+        response = requests.get(GITHUB+'cose.zip')
         with open('cose.zip', 'wb') as f:
             f.write(response.content)
         print('\nScaricato cose.zip')
@@ -79,7 +93,7 @@ try:
                         except:
                             pass
                         name, dire = line.split(';')
-                        dire = MINECRAFT+''+dire
+                        dire = MINECRAFT+dire
                         if dirr:
                             shutil.move('cosse/'+name, dire)
                         else:
@@ -90,14 +104,15 @@ try:
                 else:
                     rest = line[1:]
                     name, dire = rest.split(';')
-                    dire = MINECRAFT+''+dire
-                    try:
-                        if dirr:
-                            shutil.rmtree(dire)
-                        else:
-                            os.remove(dire+name)
-                    except:
-                        pass
+                    dire = MINECRAFT+dire
+                    if a:
+                        try:
+                            if dirr:
+                                shutil.rmtree(dire)
+                            else:
+                                os.remove(dire+name)
+                        except:
+                            pass
                     if dirr:
                         shutil.move('cosse/'+name, dire)
                     else:
@@ -108,7 +123,7 @@ try:
         
         os.remove('cose.zip')
         shutil.rmtree('cosse/')
-        if agg:
+        if agg or sc:
             tx()
         else:
             fine()
@@ -122,8 +137,15 @@ try:
             f.write(response.content)
         print('Scaricato forge.jar')
         
-        down.sc(MINECRAFT+'mods/')
-        down.scarica_file('https://github.com/dj2828/sito/releases/download/mod_vanilla_2/ultra_vanilla_2.jar', MINECRAFT+'mods/ultra_vanilla_2.jar')
+        down_error = down.sc(MINECRAFT+'mods/')
+        down.scarica_file(GITHUB+'ultra_vanilla_2.jar', MINECRAFT+'mods/ultra_vanilla_2.jar')
+
+        if down_error:
+            print("\033[91mATTENZIONE: alcune mod non sono state scaricate\033[0m")
+            print("Le mod che non sono state scaricate sono:")
+            for error in down_error:
+                print(error)
+            input('\n\033[91mPremere invio per continuare\033[0m')
 
         print("\033[92mPremere INVIO\033[0m")
         input('')
@@ -132,7 +154,7 @@ try:
 
         print('\033[92mOra comparirà una finestra per installare forge, tu prosegui')
         input('Premi INVIO per iniziare\033[0m ')
-        os.system('start '+'C:/Users/'+user+'/Downloads/forge-1.20.1-47.4.0-installer.jar')
+        os.system('start '+'C:/USERs/'+USER+'/Downloads/forge-1.20.1-47.4.0-installer.jar')
         print('\n\033[92mUna volta finito premi INVIO\033[0m')
         input('')
 
@@ -151,9 +173,16 @@ try:
 
         # scarica quelle nuove
 
-        down.upt_down()
-        
+        down_error = down.upt_down()
+
         shutil.move('./mods/', MINECRAFT+'mods/')
+
+        if down_error:
+            print("\033[91mATTENZIONE: alcune mod non sono state scaricate\033[0m")
+            print("Le mod che non sono state scaricate sono:")
+            for error in down_error:
+                print(error)
+            input('\n\033[91mPremere invio per continuare\033[0m')
 
         print('Mod aggiornate')
 
@@ -163,13 +192,13 @@ try:
         elif cos == 'n':
             tx()
 
-    print(user)
+    print(USER)
     print("Benvenuto nell' installer delle mod")
-    print("Se devi scaricare le mod scrivi 's'\nSe devi aggiornare scrivi 'a'\nSe devi aggiornare la texture pack scrivi 'tx'\nSe devi aggiornare altre cose scrivi 'cose'")
+    print("Se devi scaricare le mod scrivi 's'\nSe devi aggiornare/riparare scrivi 'a'\nSe devi aggiornare la texture pack scrivi 'tx'\nSe devi aggiornare altre cose scrivi 'cose'")
     cos = input('')
     os.system('cls')
     if cos == 's':
-        response = requests.get('https://github.com/dj2828/sito/releases/download/mod_vanilla_2/modlist.json')
+        response = requests.get(GITHUB+'modlist.json')
         with open('modlist.json', 'wb') as f:
             f.write(response.content)
         print(f'Scaricato modlist.json')
@@ -177,24 +206,32 @@ try:
         while os.path.exists(MINECRAFT+'mods'):
             print('\033[91mLa cartella mods esiste ancora, rinominala o cancellala. premi INVIO\033[0m')
             input('')
-            os.system('start '+MINECRAFT+'')
+            os.system('start '+MINECRAFT)
             print('Una volta fatto premi INVIO')
             input('')
         os.makedirs(MINECRAFT+'mods/')
 
-        agg = True
+        sc = True
         scarica_mod()
 
     elif cos == 'a':
-        if os.path.exists(MINECRAFT+'mods/') == False:
-            print("\033[91mLa cartella mods non esiste, quindi scegli 'scaricare'. premi INVIO\033[0m")
+        if not os.path.exists(MINECRAFT+'mods/modlist.json'):
+            print("\033[91mLa cartella mods non esiste o è stata rinominata, quindi scegli 'scaricare' o rinominala in 'mods'. premi INVIO\033[0m")
             input('')
+            os.system('start '+MINECRAFT)
             sys.exit()
-        response = requests.get('https://github.com/dj2828/sito/releases/download/mod_vanilla_2/modlist.json')
+        response = requests.get(GITHUB+'modlist.json')
         with open('modlist.json', 'wb') as f:
             f.write(response.content)
         print(f'Scaricato modlist.json')
         agg = True
+        if filecmp.cmp("modlist.json", MINECRAFT+'mods/modlist.json', shallow=False):
+            print("\033[92mLe mod sono già aggiornate\033[0m")
+            cos = input('\n\033[92mVuoi anche aggiornare cose? (s/n) \033[0m')
+            if cos == 's':
+                cose(True)
+            elif cos == 'n':
+                tx()
         upt_mod()
 
     elif cos == 'tx':
